feat!: Restructure library
Moves all files into `lib/src` and selectively exports classes and
functions from `lib/stroke_order_animator.dart`. This provides a cleaner
public API.
