+++ lib/src/stroke_order_animation_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:stroke_order_animator/stroke_order.dart';
+import 'package:stroke_order_animator/src/distance_2_d.dart';
 import 'package:stroke_order_animator/stroke_order_animator.dart';
 
 /// A ChangeNotifier that controls the behavior of a stroke order diagram.
@@ -553,24 +553,5 @@ class StrokeOrderAnimationController extends ChangeNotifier {
     }
 
     notifyListeners();
-  }
-}
-
-class QuizSummary {
-  QuizSummary(this._nStrokes) {
-    reset();
-  }
-  final int _nStrokes;
-  int get nStrokes => _nStrokes;
-
-  late List<int> mistakes;
-  late List<List<Offset>> correctStrokePaths;
-
-  int get nTotalMistakes =>
-      mistakes.fold(0, (previous, current) => previous + current);
-
-  void reset() {
-    mistakes = List.generate(nStrokes, (index) => 0);
-    correctStrokePaths = List.generate(nStrokes, (index) => []);
   }
 }
