+++ test/stroke_order_animator_test.dart
@@ -4,8 +4,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
-import 'package:stroke_order_animator/stroke_order.dart';
-import 'package:stroke_order_animator/stroke_order_animation_controller.dart';
+import 'package:stroke_order_animator/src/distance_2_d.dart';
 import 'package:stroke_order_animator/stroke_order_animator.dart';
 
 import 'test_strokes.dart';
