+++ test/stroke_order_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:stroke_order_animator/stroke_order.dart';
+import 'package:stroke_order_animator/stroke_order_animator.dart';
 
 import 'test_strokes.dart';
